+++ tests/settings.py
@@ -1,4 +1,0 @@
-import pathlib
-
-TESTS_ROOT = pathlib.Path(__file__).absolute().parent
-TEST_MAKEFILE = TESTS_ROOT.joinpath("data", "mk.toml")